+++ src/engine/graphic/GRFinishBar.cpp
@@ -94,7 +94,7 @@ void GRFinishBar::DrawWithLines( VGDevice & hdc ) const
         hdc.PushFillColor(VGColor(mColRef));
 
     // - Vertical adjustement according to staff's line number
-    float offsety1 = (fmod(- 0.5f * fLineNumber - 2, 3) + 1.5f) * LSPACE;
+    float offsety1 = (float)(fmod(- 0.5f * fLineNumber - 2, 3) + 1.5f) * LSPACE;
     float offsety2 = 0;
 
     if (fLineNumber != 0 && fLineNumber != 1)
